Fix uploading portrait images
A lot of tools are not able to extract and correctly interpret exif tags from jpgs. Sharp is. Also, uploading an image to tinypng, donwloading it again, then upload it to s3 took a long time. With imagemin it's possible to handle this process on server, with possibly slightly worse compression results but massive speed improvements. If this works, the socket.io dependency could be removed entirely.